Integrate Case Notes API client to replace functionality in Prison API. Refactor services, tests, and configuration to use the new API client and mock server.
