+++ helm_deploy/values-dev.yaml
@@ -14,6 +14,7 @@ generic-service:
     API_BASE_URL_OAUTH: https://sign-in-dev.hmpps.service.justice.gov.uk/auth
     API_BASE_URL_PRISON: https://prison-api-dev.prison.service.justice.gov.uk
     API_BASE_URL_PRISONER_SEARCH: https://prisoner-search-dev.prison.service.justice.gov.uk
+    API_BASE_URL_CASE_NOTES: https://dev.offender-case-notes.service.justice.gov.uk
     API_BASE_URL_LOCATIONS: https://locations-inside-prison-api-dev.hmpps.service.justice.gov.uk
     # First weekday of the month at 10:00am - Offender Search API is turned off during non-working hours in dev/preprod
     TASK_UPDATE_KPIS_CRON: "0 0 10 1W * *"
