+++ helm_deploy/values-preprod.yaml
@@ -12,6 +12,7 @@ generic-service:
     API_BASE_URL_OAUTH: https://sign-in-preprod.hmpps.service.justice.gov.uk/auth
     API_BASE_URL_PRISON: https://prison-api-preprod.prison.service.justice.gov.uk
     API_BASE_URL_PRISONER_SEARCH: https://prisoner-search-preprod.prison.service.justice.gov.uk
+    API_BASE_URL_CASE_NOTES: https://preprod.offender-case-notes.service.justice.gov.uk
     API_BASE_URL_LOCATIONS: https://locations-inside-prison-api-preprod.hmpps.service.justice.gov.uk
     # First weekday of the month at 10:00am - Offender Search API is turned off during non-working hours in dev/preprod
     TASK_UPDATE_KPIS_CRON: "0 0 10 1W * *"
