+++ helm_deploy/values-prod.yaml
@@ -10,6 +10,7 @@ generic-service:
     API_BASE_URL_OAUTH: https://sign-in.hmpps.service.justice.gov.uk/auth
     API_BASE_URL_PRISON: https://prison-api.prison.service.justice.gov.uk
     API_BASE_URL_PRISONER_SEARCH: https://prisoner-search.prison.service.justice.gov.uk
+    API_BASE_URL_CASE_NOTES: https://offender-case-notes.service.justice.gov.uk
     API_BASE_URL_LOCATIONS: https://locations-inside-prison-api.hmpps.service.justice.gov.uk
 
   postgresDatabaseRestore:
